el cliente recibe los tiles de a poco

+++ Server/Headers/Modelo/Juego.h
@@ -43,7 +43,7 @@ public:
 	bool getStatusPartida();
 	map<int,EntidadDinamica*> getProtagonistas();
 	list<EntidadDinamica*>* getNewProtagonistasToNotify();
-	pair<float,float>* getPositionOfProtagonistaById(int idBuscado);
+	pair<int,int>* getPositionOfProtagonistaById(int idBuscado);
 	void cleanNewProtagonistas();
 
 	void agregarProtagonista(string owner);
